Initial commit for parallelization

+++ src/utilities/cmc_geo_util.h
@@ -56,8 +56,8 @@ public:
     std::string name{};
 
     /* In a parallel setting, we are saving the start and count data pointer (for a blocked parallel distribution of the variable's data */
-    std::vector<uint32_t> start_ptr;
-    std::vector<uint32_t> count_ptr;
+    std::vector<uint64_t> start_ptr;
+    std::vector<uint64_t> count_ptr;
 
     void switch_data();
     void get_data_layout_from_axis_ordering();
